Add files via upload
These notes are my own and may not be entirely accurate. They are intended to help you understand the content. I apologize if the handwriting is unclear and code is difficult to follow due to handwriting. You can refer to the CODE folder in this repository, which contains the same code for better clarity.

:)